feat: add abstraction, docs, joined cryptography in one service
