+++ packages/dart/consumer_auth_provider/lib/src/provider/consumer_token_provider.dart
