+++ packages/dart/cryptography/lib/affinidi_tdk_cryptography.dart
@@ -2,5 +2,3 @@ library;
 
 export 'src/cryptography_service.dart';
 export 'src/cryptography_interface.dart';
-export 'src/aes_cbc_encryption_service.dart';
-export 'src/aes_cbc_encryption_interface.dart';
